line_edit: Implement initial version
Command line editor is triggered when the key 'e' is pressed. When
editing the command line, the key ranging from 27 to 127 (printable
ASCII characters) are included in the options. Backspace delete the
character before the cursor and Enter confirm the changes.

Supports the keys: ESC (cancel), Home (goto begin), End (goto end),
Delete (delete character on the cursor) and the left and right arrows.